requirements.txt
yfinance
requests
feedparser
openai
python-dotenv
pyyaml
